+++ src/main/scala/org/clulab/wm/eidos/refiners/EidosRefiner.scala
@@ -1,7 +1,7 @@
 package org.clulab.wm.eidos.refiners
 
-import org.clulab.wm.eidos.EidosComponents
 import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.components.EidosComponents
 import org.clulab.wm.eidos.document.AnnotatedDocument
 import org.clulab.wm.eidos.utils.Timer
