+++ src/main/scala/org/clulab/wm/eidos/refiners/OdinRefiner.scala
@@ -1,8 +1,8 @@
 package org.clulab.wm.eidos.refiners
 
 import org.clulab.odin.Mention
-import org.clulab.wm.eidos.EidosComponents
 import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.components.EidosComponents
 import org.clulab.wm.eidos.utils.Timer
 
 class OdinRefiner(val name: String, val refine: Seq[Mention] => Option[Seq[Mention]])
