+++ src/main/scala/org/clulab/wm/eidos/refiners/ProcessorRefiner.scala
@@ -1,8 +1,8 @@
 package org.clulab.wm.eidos.refiners
 
 import org.clulab.processors.Document
-import org.clulab.wm.eidos.EidosComponents
 import org.clulab.wm.eidos.EidosSystem
+import org.clulab.wm.eidos.components.EidosComponents
 import org.clulab.wm.eidos.document.Metadata
 import org.clulab.wm.eidos.document.attachments.RelevanceDocumentAttachment
 import org.clulab.wm.eidos.utils.Timer
